Refactored Sidebar and its modals
- created two new folders, one for the components and one for the item on the page, its beautiful
